reverting before the flocking behaviours where changed from enemies to flockCubes,

+++ AI/aieBootstrap-master/project2D/Enemy.h
@@ -17,7 +17,7 @@ public:
 	Enemy();
 	Enemy(float positionX, float positionY);
 	~Enemy();
-	int health;
+	double health;
 	void render();
 	void update(float dt);
 	void changeToSeek();
